Fix initscript JAVA and JAVA_HOME variables override.
Configuring JAVA and JAVA_HOME variables override on
/etc/default/logstash (DEBIAN) or /etc/sysconfig/logstash (RHEL)
Was working only on Debian initscript.

The problem was that JAVA variable was configured on redhat initscript
after the sysconfig/logstash file was loaded.

This commit fix this problem and now it is possible to configure JAVA
and JAVA_HOME on RHEL and on Debian.